+++ Scripts/integer/integer2.py
@@ -1,1 +1,1 @@
-print(float(input()) // 1000)
+print(int(input()) // 1000)
